correct grammar in quickstart
https://github.com/SSSD/sssd.io/blob/master/src/docs/quick-start.rst

As your intent is likely for people to use the quickstart guide as a template to start their own sssd installation, you should change the **i.e.**  to  **e.g.**.

i.e. would imply that there is only one possible setting. And as the end user does not own the domain specified (sssd.io) that would contradict the documentation.

e.g. provides an example which is following in the worked example below

Fixes #62